requirements.txt
telethon==1.36.0
google-generativeai==0.7.2
requests==2.32.3
python-dotenv==1.0.1
